CS454: Theory of Computation
This project is attempting to solve an arbitrary problem using DFAs. It was something like binary with reversing negatives e.g. instead of 
0111 = (0*2^3) + (1*2^2) + (1*2^1) + (1*2^0) = 7
0111 = (0*2^3) + (1*2^2) - (1*2^1) + (1*2^0) = 3

I don't think I was 100% successful, but I still got an A. Nobody has solved this problem in the history of the class, including the professor. He only solved it theoritcally.